+++ paper.bib
@@ -12,16 +12,22 @@
 }
 
 @article{arakawa_computational_1966,
-  title   = {Computational design for long-term numerical integration of the equations of fluid motion: {Two}-dimensional incompressible flow. {Part} {I}},
-  volume  = {1},
-  issn    = {0021-9991},
-  url     = {https://www.sciencedirect.com/science/article/pii/0021999166900155},
-  doi     = {https://doi.org/10.1016/0021-9991(66)90015-5},
-  number  = {1},
-  journal = {Journal of Computational Physics},
-  author  = {Arakawa, Akio},
-  year    = {1966},
-  pages   = {119--143}
+  author     = {Arakawa, Akio},
+  title      = {Computational Design for Long-Term Numerical Integration of the Equations of Fluid Motion},
+  year       = {1997},
+  issue_date = {Aug. 1997},
+  publisher  = {Academic Press Professional, Inc.},
+  address    = {USA},
+  volume     = {135},
+  number     = {2},
+  issn       = {0021-9991},
+  url        = {https://doi.org/10.1006/jcph.1997.5697},
+  doi        = {10.1006/jcph.1997.5697},
+  abstract   = {The integral constraints on quadratic quantities of physical importance, such as conservation of mean kinetic energy and mean square vorticity, will not be maintained in finite difference analogues of the equation of motion for two-dimensional incompressible flow, unless the finite difference Jacobian expression for the advection term is restricted to a form which properly represents the interaction between grid points, as derived in this paper. It is shown that the derived form of the finite difference Jacobian prevents nonlinear computational instability and thereby permits long-term numerical integrations.},
+  journal    = {J. Comput. Phys.},
+  month      = {aug},
+  pages      = {103–114},
+  numpages   = {12}
 }
 
 @article{grillix,
@@ -53,13 +59,19 @@
 }
 
 @article{anderson2020elucidating,
-  title     = {Elucidating plasma dynamics in Hasegawa--Wakatani turbulence by information geometry},
-  author    = {Anderson, Johan and Kim, Eun-jin and Hnat, Bogdan and Rafiq, Tariq},
-  journal   = {Physics of Plasmas},
-  volume    = {27},
-  number    = {2},
-  year      = {2020},
-  publisher = {AIP Publishing}
+  author   = {Anderson, Johan and Kim, Eun-jin and Hnat, Bogdan and Rafiq, Tariq},
+  title    = {{Elucidating plasma dynamics in Hasegawa–Wakatani turbulence by information geometry}},
+  journal  = {Physics of Plasmas},
+  volume   = {27},
+  number   = {2},
+  pages    = {022307},
+  year     = {2020},
+  month    = {02},
+  abstract = {{The impact of adiabatic electrons on drift-wave turbulence, modeled by the Hasegawa–Wakatani equations, is studied using information length. Information length is a novel theoretical method for measuring distances between statistical states represented by different probability distribution functions (PDFs) along the path of a system and represents the total number of statistically different states that a system evolves through in time. Specifically, the time-dependent PDFs of turbulent fluctuations for a given adiabatic index A are computed. The changes in fluctuation statistics are then quantified in time by using information length. The numerical results provide time traces exhibiting intermittent plasma dynamics, and such behavior is identified by a rapid change in the information length. The effects of A are discussed.}},
+  issn     = {1070-664X},
+  doi      = {10.1063/1.5122865},
+  url      = {https://doi.org/10.1063/1.5122865},
+  eprint   = {https://pubs.aip.org/aip/pop/article-pdf/doi/10.1063/1.5122865/15771314/022307\_1\_online.pdf}
 }
 
 @article{hasegawa_wakatani,
[…]
