### 2023-08-21 #5773 @kyleniemeyer
Hi @the-rccg, thanks for your submission to JOSS. We will use this pre-review issue to assign an editor and find reviewers. Unfortunately, we do not have an editor who is available to handle your submission right now, so I have to place this on our waitlist until someone becomes available.

In the meantime, reviewer suggestions are welcome!
