### 2023-08-21 #5773 @the-rccg
@kyleniemeyer thank you very much for opening the pre-review! I'll looking over the reviewer list, it might be useful to broaden the domain since plasma turbulence is a interdisciplinary field including astrophysics. Maybe an editor from that field is available?
