### 2023-08-21 #5773 @the-rccg
I did a quick search found a few potential matches, maybe some of them are willing to give it a look
- [Yehor Yudin](https://reviewers.joss.theoj.org/reviewers/3064) works on very similar topics in plasma physics and I'm familiar with his work in the field
- [Uddipta Bhardwaj](https://reviewers.joss.theoj.org/reviewers/3075) works with simulation based inference on astrophysical models which is close to the application and domain
- [Thomas A. Marks](https://reviewers.joss.theoj.org/reviewers/2904) seems to work on fluid plasma physics too, although very different aspects for thrusters rathe r than fusion
- [olgadoronina](https://reviewers.joss.theoj.org/reviewers/1943) lists turbulence modeling as a topic, but not much on the JOSS page to go off
- [Ashwin Vishnu Mohanan](https://reviewers.joss.theoj.org/reviewers/613) lists CFD turbulence and Python
- [Hongyang Zhou](https://reviewers.joss.theoj.org/reviewers/2620) seems to work on related topics in Helsinki, but sadly does not have Python listed as a programming language
