### 2023-08-24 bf4f871
Fixing of typos in readme
