### 2023-10-18 #5959 @editorialbot
[REVIEW]: HW2D: A reference implementation of the Hasegawa-Wakatani model for plasma turbulence in fusion reactors
**Submitting author:** @the-rccg (<a href="http://orcid.org/0000-0003-4143-780X">Robin Greif</a>)
**Repository:** https://github.com/the-rccg/hw2d
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.0
