### 2023-10-25 0e4df3f
WIP - Fix L2-norm with square inside the sum.
