### 2023-10-25 b28db4c
WIP - Rename variables that now include the squared L2-Norm.
