### 2023-10-31 #5959 @archermarx
@Kevin-Mattheus-Moerman I have completed most of my review, except for actually running the software. Overall, I think this is a very nice contribution. I have two main comments:

1. You do a good job of documenting the available methods, but you are missing an example or two to demonstrate how to use the code. I would include a small tutorial in the documentation showing how to use the code and showing off a few configurable options.

2. On a similar point, you mention in the installation instructions that you can use accelerators like numba. It would be good to have a list of which ones are supported by the code.
