### 2023-11-10 2394aa5
Add Example usage for CLI with indepth explanation
