### 2023-11-10 #5959 @the-rccg
@archermarx I added an example for the CLI interface with an explanation for the kwargs used. There is also an additional paragraph on the accelerators that clearly labels available accelerators (currently only Numba). Others should be ready soon, and will be added to the readme and installation. For a full workflow example, a more thorough framework would be needed like a Google Colab Notebook, which is planned for the future.
