### 2023-11-10 #5959 @archermarx
Great! I've now run the code and it mostly works. However, I encounter an error using the following input

```python3 -m hw2d --step_size=0.025 --end_time=100 --grid_pts=128 --c1=1.0 --k0=0.15 --N=3 --nu=5.0e-08 --snaps=1 --buffer_size=100 --output_path="test.h5" --movie=1 --min_fps=10 --speed=5 --debug=0```

The code runs to completion and generates a (very nice looking) movie. However, in the "Plotting properties" step, I get this error:

```
100%|███████████████████████████████████████████████████████████████████████████████████| 41/41 [00:16<00:00,  2.46it/s]
Plotting properties...███████████████████████████████████████████████████████████████▉  | 40/41 [00:16<00:00,  2.53it/s]
/home/marksta/.local/lib/python3.10/site-packages/numpy/core/_methods.py:206: RuntimeWarning: Degrees of freedom <= 0 for slice
  ret = _var(a, axis=axis, dtype=dtype, out=out, ddof=ddof,
/home/marksta/.local/lib/python3.10/site-packages/numpy/core/_methods.py:163: RuntimeWarning: invalid value encountered in divide
  arrmean = um.true_divide(arrmean, div, out=arrmean,
/home/marksta/.local/lib/python3.10/site-packages/numpy/core/_methods.py:198: RuntimeWarning: invalid value encountered in divide
  ret = ret.dtype.type(ret / rcount)
/home/marksta/.local/lib/python3.10/site-packages/numpy/core/fromnumeric.py:3504: RuntimeWarning: Mean of empty slice.
  return _methods._mean(a, axis=axis, dtype=dtype,
/home/marksta/.local/lib/python3.10/site-packages/numpy/core/_methods.py:129: RuntimeWarning: invalid value encountered in divide
  ret = ret.dtype.type(ret / rcount)
test_enstrophy-energy-kinetic_energy-thermal_energy.jpg
202it [02:02,  1.65it/s]
```

And no image is produced.
