### 2023-11-10 fc25ddc
Prevent some crashes, incl. proper warning messages
