### 2023-11-10 #5959 @the-rccg
@archermarx I updated the error handling to give graceful explanations and handling of this case. The issue was that end_time=100 is not inside of the saturated turbulent phase, so all statistical properties would be meaningless for the turbulent system. It now provides a warning and includes a picture of the non-converged data.
