### 2023-11-25 #5959 @Uddiptaatwork
I have now concluded my review. The code works great and the tests ran smoothly without any errors.
Here is one point that the author might consider:
- The documentation could use a home/introduction page very similar to the github-repo README. It would help new users accessing the code directly via the documentation page to get an overview of what it entails.

The documentation pertaining to the API is very well done and appears highly useful for any new user of the code. The article is succinct with analytical descriptions and citations where necessary.

Overall, I would like to congratulate the author for a very well written code and article.
I can mark my review as completed.
