### 2023-12-01 980c2bb
Fix Namespace behavior in subtraction
