### 2023-12-08 #5959 @kyleniemeyer
@the-rccg in the past, I think I converted my Markdown README to ReST and then used that for the docs generation to avoid some issues like that, but may have been with Sphinx. This looks like a good solution, though
