### 2023-12-11 466ca57
move to pdoc and remove alignments for docs
