+++ .github/workflows/pdoc.yml
@@ -60,4 +60,4 @@ jobs:
         github_token: ${{ secrets.GH_PAT }}
         publish_dir: ./docs
         publish_branch: gh-pages 
-        keep_files: true
+        keep_files: false
