+++ paper.bib
@@ -1,108 +1,0 @@
-@article{camargo,
-  author  = {Camargo, Suzana J. and Biskamp, Dieter and Scott, Bruce D.},
-  title   = {Resistive drift-wave turbulence},
-  journal = {Physics of Plasmas},
-  volume  = {2},
-  number  = {1},
-  pages   = {48-62},
-  year    = {1995},
-  doi     = {10.1063/1.871116},
-  url     = {https://doi.org/10.1063/1.871116},
-  eprint  = {https://doi.org/10.1063/1.871116}
-}
-
-@article{arakawa_computational_1966,
-  title   = {Computational design for long-term numerical integration of the equations of fluid motion: {Two}-dimensional incompressible flow. {Part} {I}},
-  volume  = {1},
-  issn    = {0021-9991},
-  url     = {https://www.sciencedirect.com/science/article/pii/0021999166900155},
-  doi     = {https://doi.org/10.1016/0021-9991(66)90015-5},
-  number  = {1},
-  journal = {Journal of Computational Physics},
-  author  = {Arakawa, Akio},
-  year    = {1966},
-  pages   = {119--143}
-}
-
-@article{grillix,
-  doi       = {10.1088/1361-6587/aaa373},
-  url       = {https://doi.org/10.1088/1361-6587/aaa373},
-  year      = 2018,
-  month     = {jan},
-  publisher = {{IOP} Publishing},
-  volume    = {60},
-  number    = {3},
-  pages     = {035005},
-  author    = {Andreas Stegmeir and David Coster and Alexander Ross and Omar Maj and Karl Lackner and Emanuele Poli},
-  title     = {{GRILLIX}: a {3D} turbulence code based on the flux-coordinate independent approach},
-  journal   = {Plasma Physics and Controlled Fusion}
-}
-
-@article{Yatomi_2023,
-  doi       = {10.1088/1361-6587/ace993},
-  url       = {https://dx.doi.org/10.1088/1361-6587/ace993},
-  year      = {2023},
-  month     = {aug},
-  publisher = {IOP Publishing},
-  volume    = {65},
-  number    = {9},
-  pages     = {095014},
-  author    = {Go Yatomi and Motoki Nakata and Makoto Sasaki},
-  title     = {Data-driven modal analysis of nonlinear quantities in turbulent plasmas using multi-field singular value decomposition},
-  journal   = {Plasma Physics and Controlled Fusion}
-}
-
-@article{anderson2020elucidating,
-  title     = {Elucidating plasma dynamics in Hasegawa--Wakatani turbulence by information geometry},
-  author    = {Anderson, Johan and Kim, Eun-jin and Hnat, Bogdan and Rafiq, Tariq},
-  journal   = {Physics of Plasmas},
-  volume    = {27},
-  number    = {2},
-  year      = {2020},
-  publisher = {AIP Publishing}
-}
-
-@article{hasegawa_wakatani,
-  title     = {Plasma edge turbulence},
-  author    = {Hasegawa, Akira and Wakatani, Masahiro},
-  journal   = {Physical Review Letters},
-  volume    = {50},
-  number    = {9},
-  pages     = {682},
-  year      = {1983},
-  publisher = {APS}
-}
-
-@article{goumiri2013reduced,
-  title     = {Reduced-order model based feedback control of the modified Hasegawa-Wakatani model},
-  author    = {Goumiri, IR and Rowley, Clarence W and Ma, Z and Gates, DA and Krommes, JA and Parker, JB},
-  journal   = {Physics of Plasmas},
-  volume    = {20},
-  number    = {4},
-  year      = {2013},
-  publisher = {AIP Publishing}
-}
-
[…]
