+++ paper.md
@@ -1,140 +1,0 @@
----
-title: 'HW2D: A reference implementation of the Hasegawa-Wakatani model for plasma turbulence in fusion reactors'
-tags:
-  - Python
-  - plasma physics
-  - dynamics
-  - fluid dynamics
-  - simulation
-authors:
-  - name: Robin Greif
-    orcid: 0000-0002-3957-2474
-    equal-contrib: true
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
-affiliations:
- - name: Max-Planck Institute for Plasma Physics, Germany
-   index: 1
- - name: Technical University Munich, Germany
-   index: 2
-date: 13 August 2023
-bibliography: paper.bib
-
----
-
-# Summary
-
-`HW2D` is an open source reference implementation of the Hasegawa-Wakatani model
-against which new methods are benchmarked, validated, and trained upon. 
-The model has become the de-facto standard model for testing new methods on
-turbulence in fusion devices. Besides providing consistent numerical simulations, 
-it implements the physical property calculations with the normalizations 
-common in reference texts that are often not clearly understandable outside of the 
-community. Along with it, an iteratively expanding list of stable parameters are 
-published for simulations with associated physical properties. This drastically reduces
-the resource usage of grid searches for stable turbulent simulations of a complex system
-and the time needed to verify implementation for testing new methods.
-
-
-# Statement of need
-
-Nuclear fusion provides the promise of clean, endless energy for humanity
-in the future. One of the biggest open questions standing in the way of 
-commercially viable fusion devices is confining the energy within the plasma that 
-is transported out through drift wave turbulence. However, the immense complexity 
-of modeling turbulence inside potential reactors still remains outside the realm 
-of the biggest supercomputers. Hence, the development of new methods for simulating 
-plasma turbulence is a pressing issue with the speed at which new methods are being 
-tested expanding rapidly. 
-
-The Hasegawa-Wakatani (HW) model has become the de-facto standard for testing methods,
-especially those developed for Navier-Stokes (NS) fluids, for fusion turbulence due 
-to its similarity to it [@camargo]. Although first published 40 years ago [@hasegawa_wakatani], 
-it is still missing an open-source reference implementation. With its rising popularity 
-in other fields, especially machine learning to test limits of physics-based learning 
-methods, a simple implementation is needed. A unified reference allows comparing methods 
-[@Yatomi_2023;@grillix;@anderson2020elucidating;@goumiri2013reduced;@heinonen2020learning]
-and reduces the overhead involved in verifying turbulent system implementations.
-This verification is now automated with CI pipelines and open-sourced for the broader 
-community to use.
-
-In summary, `HW2D` provides an easy to understand, easy to use, and easy to extend 
-reference implementation to help accelerate the testing of new ideas and methods.
-This stands in contrast to the closed-source Fortran and C++ codes available today at 
-a select few research institutes. As an open-source Python reference solver, 
-`HW2D` opens up the research field beyond these confines. It significantly decreases the 
-barrier of entry for fusion- and machine learning reserchers to tackle one of the 
-biggest open questions in contemporary science. This implementation was and is being 
-used for scientific papers on data-driven methods for plasma turbulence and 
[…]
