### 2023-12-11 #5959 @the-rccg
@kyleniemeyer After a lot of back and forth, I have now moved from pdoc3 to pdoc. Images and equations now work fully dynamically.
