+++ README.md
@@ -70,11 +70,11 @@ $$
                      - \left[ \phi, n \right]
                      - \kappa_n \partial_y \phi
                      - \nu \nabla^{2N} n 
-    \\
+    \\\
     \partial_t \Omega &= c_1 \left( \phi - n \right)
                                       - \left[ \phi, \Omega \right]
                                       - \nu \nabla^{2N} \Omega 
-    \\
+    \\\
     \Omega &= \nabla^2 \phi
 \end{align}
 $$
@@ -106,9 +106,9 @@ This allows simple comparison, as well es evalutaion of new methods to one refer
 
 $$
 \begin{align}
-    \Gamma^n &= -     \iint{ \mathrm{d}^2x \space \left( n \space \partial_y \phi \right) }  \\
-    \Gamma^c &= c_1   \iint{ \mathrm{d}^2x \space \left(n - \phi \right)^2}  \\
-    E        &= \small \frac{1}{2} \normalsize \iint{\mathrm{d}^2 x \space \left(n^2 - \left|\nabla_\bot \phi \right|^2 \right)}  \\
+    \Gamma^n &= -     \iint{ \mathrm{d}^2x \space \left( n \space \partial_y \phi \right) }  \\\
+    \Gamma^c &= c_1   \iint{ \mathrm{d}^2x \space \left(n - \phi \right)^2}  \\\
+    E        &= \small \frac{1}{2} \normalsize \iint{\mathrm{d}^2 x \space \left(n^2 - \left|\nabla_\bot \phi \right|^2 \right)}  \\\
     U        &= \small \frac{1}{2} \normalsize \iint{\mathrm{d}^2 x \space \left(n-\nabla_\bot^2  \phi\right)^2} = \small \frac{1}{2} \normalsize \iint{\mathrm{d}^2 x \space \left(n-\Omega\right)^2}    
 \end{align}
 $$
@@ -120,9 +120,9 @@ Additionally, some spectral properties are included for more detailed analysis b
 
 $$
 \begin{align}
-  \Gamma^n \small (k_y) \normalsize  &= - i k_y  \space n \small (k_y) \normalsize \space \phi^* \small (k_y) \normalsize  \\
-  \delta \small (k_y) \normalsize    &= - \mathrm{Im}\left( \mathrm{log} \left( n^* \small (k_y) \normalsize \space \phi \small (k_y) \normalsize \right) \right)   \\
-  E^N \small (k_y) \normalsize       &= \small \frac{1}{2}\normalsize \big| n \small (k_y) \normalsize \big|^2   \\
+  \Gamma^n \small (k_y) \normalsize  &= - i k_y  \space n \small (k_y) \normalsize \space \phi^* \small (k_y) \normalsize  \\\
+  \delta \small (k_y) \normalsize    &= - \mathrm{Im}\left( \mathrm{log} \left( n^* \small (k_y) \normalsize \space \phi \small (k_y) \normalsize \right) \right)   \\\
+  E^N \small (k_y) \normalsize       &= \small \frac{1}{2}\normalsize \big| n \small (k_y) \normalsize \big|^2   \\\
   E^V \small (k_y) \normalsize       &= \small \frac{1}{2}\normalsize \big| k_y \space \phi \small (k_y) \normalsize \big|^2     
 \end{align}
 $$
@@ -136,10 +136,10 @@ The definitions are given by:
 
 $$
 \begin{align}
[…]
