### 2023-12-12 #5959 @the-rccg
Additional Author Tasks After Review is Complete

[x] Double check authors and affiliations (including ORCIDs)
[x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
[x] Archive the release on Zenodo/figshare/etc and post the DOI here.
[x] Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
[x] Make sure that the license listed for the archive is the same as the software license.

Release: v1.0.0 (Official JOSS release)
- Check GitHub: https://github.com/the-rccg/hw2d/releases/tag/release
- Check PyPI: https://pypi.org/project/hw2d/1.0.0/
ORCID from Zenodo: https://zenodo.org/doi/10.5281/zenodo.10365012
