### 2023-12-12 #5959 @kyleniemeyer
@the-rccg I made a few edits to the paper, can you merge those? https://github.com/the-rccg/hw2d/pull/4
