### 2023-12-12 #5959 @kyleniemeyer
Also, can you add the DOI to the Goumiri et al. reference? 10.1063/1.4796190
