### 2023-12-12 1fc28f5
Add remainign DOIs to bibliography
