+++ paper.bib
@@ -1,4 +1,5 @@
 @article{camargo,
+  doi     = {10.1063/1.871116},
   author  = {Camargo, Suzana J. and Biskamp, Dieter and Scott, Bruce D.},
   title   = {Resistive drift-wave turbulence},
   journal = {Physics of Plasmas},
@@ -6,12 +7,12 @@
   number  = {1},
   pages   = {48-62},
   year    = {1995},
-  doi     = {10.1063/1.871116},
   url     = {https://doi.org/10.1063/1.871116},
   eprint  = {https://doi.org/10.1063/1.871116}
 }
 
 @article{arakawa_computational_1966,
+  doi        = {10.1006/jcph.1997.5697},
   author     = {Arakawa, Akio},
   title      = {Computational Design for Long-Term Numerical Integration of the Equations of Fluid Motion},
   year       = {1997},
@@ -22,7 +23,6 @@
   number     = {2},
   issn       = {0021-9991},
   url        = {https://doi.org/10.1006/jcph.1997.5697},
-  doi        = {10.1006/jcph.1997.5697},
   abstract   = {The integral constraints on quadratic quantities of physical importance, such as conservation of mean kinetic energy and mean square vorticity, will not be maintained in finite difference analogues of the equation of motion for two-dimensional incompressible flow, unless the finite difference Jacobian expression for the advection term is restricted to a form which properly represents the interaction between grid points, as derived in this paper. It is shown that the derived form of the finite difference Jacobian prevents nonlinear computational instability and thereby permits long-term numerical integrations.},
   journal    = {J. Comput. Phys.},
   month      = {aug},
@@ -59,6 +59,7 @@
 }
 
 @article{anderson2020elucidating,
+  doi      = {10.1063/1.5122865},
   author   = {Anderson, Johan and Kim, Eun-jin and Hnat, Bogdan and Rafiq, Tariq},
   title    = {{Elucidating plasma dynamics in Hasegawa–Wakatani turbulence by information geometry}},
   journal  = {Physics of Plasmas},
@@ -69,12 +70,12 @@
   month    = {02},
   abstract = {{The impact of adiabatic electrons on drift-wave turbulence, modeled by the Hasegawa–Wakatani equations, is studied using information length. Information length is a novel theoretical method for measuring distances between statistical states represented by different probability distribution functions (PDFs) along the path of a system and represents the total number of statistically different states that a system evolves through in time. Specifically, the time-dependent PDFs of turbulent fluctuations for a given adiabatic index A are computed. The changes in fluctuation statistics are then quantified in time by using information length. The numerical results provide time traces exhibiting intermittent plasma dynamics, and such behavior is identified by a rapid change in the information length. The effects of A are discussed.}},
   issn     = {1070-664X},
-  doi      = {10.1063/1.5122865},
   url      = {https://doi.org/10.1063/1.5122865},
   eprint   = {https://pubs.aip.org/aip/pop/article-pdf/doi/10.1063/1.5122865/15771314/022307\_1\_online.pdf}
 }
 
 @article{hasegawa_wakatani,
+  doi       = {10.1103/PhysRevLett.50.682}, 
   title     = {Plasma edge turbulence},
   author    = {Hasegawa, Akira and Wakatani, Masahiro},
   journal   = {Physical Review Letters},
@@ -86,6 +87,7 @@
 }
 
 @article{goumiri2013reduced,
+  doi       = {10.1063/1.4796190},
   title     = {Reduced-order model based feedback control of the modified Hasegawa-Wakatani model},
[…]
