### 2023-12-12 #5959 @the-rccg
Can the Tex warnings be safely ignored or do I need to removethe \space and \small from it still?
