### 2023-12-12 #5959 @kyleniemeyer
@the-rccg let's try removing `\space` and `\small` in the equations. If you need spacing in math mode, I believe you can use `\,`, `\:`, and `\;` instead.

As for `\small`, I'm not sure the output you were aiming for there. Is that needed?
