### 2023-12-12 490cc99
cleanup paper equations for JOSS
