### 2023-12-12 #5959 @the-rccg
@kyleniemeyer removed all \small, \normalsize, and \space and replaces some with more explicit spacings.
