### 2023-12-12 c09e9b1
Add CITATION.cff
